SynthesizerContext cleanup
